requirements.txt
Flask
Flask-SQLAlchemy
Pillow
google-cloud-vision
python-dotenv
clip-interrogator
torch
transformers
sentence-transformers
accelerate
